Revert "Calculator alternative"
